chore: Rename `fetchAllEntriesFrom` to `getAllStoredEntriesFromSender`
